drop upperbounds and bump version
The library is not using anything other than the most basic functions from each of its dependencies (namely, fromList and toList) which I expect won't ever change in the future, so upperbounds aren't needed.

If upper bounds will be needed in the future, the bounds can still be retroactively fixed on hackage directly.

Closes #8